# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /studio/images:
    post:
      operationId: generateStudioImage
      tags: [studio]
      summary: Generate a reference-guided scene image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioImageInput"
      responses:
        "200":
          description: Generated image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioImage"
        "400":
          description: Invalid generation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Image provider could not complete the request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /studio/session:
    get:
      operationId: getStudioSession
      tags: [studio]
      summary: Report whether the caller has an unlocked operator session
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioSessionStatus"
    post:
      operationId: createStudioSession
      tags: [studio]
      summary: Unlock the studio with the operator access password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioSessionInput"
      responses:
        "200":
          description: Studio unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioSessionStatus"
        "401":
          description: Incorrect access password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudioSession
      tags: [studio]
      summary: Lock the studio (clear the operator session)
      responses:
        "204":
          description: Studio locked
  /studio/scenes:
    get:
      operationId: getStudioScenes
      tags: [studio]
      summary: List generated studio scenes
      responses:
        "200":
          description: Generated scenes, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudioScene"
        "401":
          description: Studio is locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createStudioScene
      tags: [studio]
      summary: Save a generated studio scene
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioSceneInput"
      responses:
        "201":
          description: Saved generated scene
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioScene"
        "400":
          description: Invalid scene
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Studio is locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Scene image is too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudioScenes
      tags: [studio]
      summary: Delete all generated studio scenes
      responses:
        "204":
          description: All generated scenes deleted
        "401":
          description: Studio is locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /studio/scenes/{sceneId}:
    delete:
      operationId: deleteStudioScene
      tags: [studio]
      summary: Delete one generated studio scene
      parameters:
        - name: sceneId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Generated scene deleted
        "401":
          description: Studio is locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Generated scene not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /studio/post-copy:
    post:
      operationId: generateStudioPostCopy
      tags: [studio]
      summary: Generate an editable Instagram caption for a studio asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioPostCopyInput"
      responses:
        "200":
          description: Generated post copy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioPostCopy"
        "400":
          description: Invalid post-copy request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Copy provider could not complete the request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /studio/videos:
    post:
      operationId: startStudioVideo
      tags: [studio]
      summary: Start a face-preserving Reel or Story video render
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioVideoInput"
      responses:
        "200":
          description: Video render task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioVideoTask"
        "400":
          description: Invalid video render request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A public deployment URL is required for video rendering
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Video rendering is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Video provider could not start the render
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /studio/videos/{taskId}:
    get:
      operationId: getStudioVideo
      tags: [studio]
      summary: Get the current status and result of a studio video render
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current video render task state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioVideoTask"
        "404":
          description: Video render task was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Video provider could not retrieve the render
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /studio/assets/{assetId}:
    get:
      operationId: getStudioAsset
      tags: [studio]
      summary: Retrieve a temporary image asset prepared for publishing
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image asset
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        "404":
          description: Asset not found or expired
  /studio/instagram/connect:
    post:
      operationId: beginInstagramConnection
      tags: [studio]
      summary: Start the Instagram account connection flow
      responses:
        "200":
          description: Instagram authorization link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramConnection"
        "503":
          description: Instagram publishing is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /studio/instagram/publish:
    post:
      operationId: publishStudioImageToInstagram
      tags: [studio]
      summary: Publish a generated image to the connected Instagram account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstagramPublishInput"
      responses:
        "200":
          description: Published Instagram post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstagramPublication"
        "400":
          description: Invalid publishing request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A public deployment URL is needed before publishing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Instagram could not complete the post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    StudioImageInput:
      type: object
      required: [prompt, aspectRatio]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 6000
        aspectRatio:
          type: string
          enum: ["16:9", "9:16", "1:1"]
        referenceImage:
          type: ["string", "null"]
          maxLength: 16000000
        fidelity:
          type: string
          enum: ["high", "balanced"]
          default: high
    StudioImage:
      type: object
      required: [imageDataUrl, provider, referenceUsed, fidelity]
      properties:
        imageDataUrl:
          type: string
        provider:
          type: string
        referenceUsed:
          type: boolean
        fidelity:
          type: string
          enum: ["high", "balanced"]
    StudioSceneInput:
      type: object
      required: [prompt, aspectRatio, fidelity, referenceUsed, imageDataUrl, provider]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 6000
        aspectRatio:
          type: string
          enum: ["16:9", "9:16", "1:1"]
        fidelity:
          type: string
          enum: ["high", "balanced"]
        referenceUsed:
          type: boolean
        imageDataUrl:
          type: string
          maxLength: 16000000
        provider:
          type: string
          maxLength: 200
    StudioScene:
      type: object
      required: [id, prompt, aspectRatio, fidelity, referenceUsed, imageDataUrl, provider, createdAt]
      properties:
        id:
          type: string
        prompt:
          type: string
        aspectRatio:
          type: string
          enum: ["16:9", "9:16", "1:1"]
        fidelity:
          type: string
          enum: ["high", "balanced"]
        referenceUsed:
          type: boolean
        imageDataUrl:
          type: string
        provider:
          type: string
        createdAt:
          type: string
          format: date-time
    StudioSessionInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 200
    StudioSessionStatus:
      type: object
      required: [unlocked, required]
      properties:
        unlocked:
          type: boolean
          description: Whether the caller currently holds a valid operator session
        required:
          type: boolean
          description: Whether an access password is configured (false in open dev mode)
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string